agents/frameworks/agno/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibmi-agents"
version = "0.1.0"
description = "IBM i specialized agents with MCP tools and comprehensive evaluation framework"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "IBM i MCP Team"}
]
license = {text = "MIT"}
keywords = ["ibm-i", "agents", "mcp", "system-administration", "agno"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "anthropic>=0.67.0",
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.12.2",
    "mcp>=1.13.1",
    "openai>=1.107.1",
    "sqlalchemy>=2.0.43",
    "toolbox-core>=0.5.0",
    "uvicorn>=0.35.0",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "ibm-watsonx-ai>=1.3.39",
    "ollama>=0.6.0",
    "agno>=2.0.11",
    "ibmi-agent-sdk",
]

[tool.uv.sources]
ibmi-agent-sdk = { path = "../../packages/ibmi-agent-sdk", editable = true }
